+++ lib/rails_stuff/resources_controller/sti_helpers.rb
@@ -4,6 +4,8 @@ module RailsStuff
   module ResourcesController
     # Helper methods for controllers which works with STI models.
     module StiHelpers
+      class InvalidType < StandardError; end
+
       extend ActiveSupport::Concern
 
       module ClassMethods
@@ -12,14 +14,27 @@ module RailsStuff
         # By default it uses `.types_list` or `.descendants` to get list of
         # classes and indexes them by class names.
         def resource_class_by_type
-          @resource_class_by_type ||= begin
+          @resource_class_by_type ||=
             if resource_class.respond_to?(:types_list)
               resource_class.types_list
             else
               resource_class.descendants
             end.index_by(&:name)
-          end
         end
+
+        attr_writer :resource_class_by_type
+
+        def resource_class_for(name)
+          return resource_class unless name
+          resource_class_by_type[name] ||
+            if use_resource_class_for_invalid_type
+              resource_class
+            else
+              raise(InvalidType, "No type mapped for #{name.inspect}")
+            end
+        end
+
+        attr_accessor :use_resource_class_for_invalid_type
 
         # Class-level accessor to permitted attributes for specisic class.
         def permitted_attrs_for
@@ -41,15 +56,14 @@ module RailsStuff
           if params.key?(:id)
             resource.class
           else
-            name = params.require(self.class.resource_param_name).
-              permit(:type)[:type]
-            self.class.resource_class_by_type[name] ||
-              raise(ActiveRecord::RecordNotFound)
+            key = self.class.resource_param_name
+            name = params.permit(key => [:type])[key].try!(:[], :type)
+            self.class.resource_class_for(name)
           end
       end
 
       # Instantiates object using class_from_request.
-      def build_resource
+      def build_resource(attrs = resource_params)
         @_resource = super.becomes!(class_from_request)
       end